+++ CHANGELOG.md
@@ -41,3 +41,4 @@
 - 'EmDuration' added 'to' method to get duration days, hours, minutes, seconds and milliseconds.
 - Added new 'EmTag' & 'EmTags' classes to have simple synchronization among same values coming from different sources
 - Added new 'EmAutoPtr' class as a lightweight alternative to std::auto_ptr for environments where the standard library is not available or desired
+- Removed 'SoftwareSerial' as dependency since its not more used.